client: store topcolor and bottomcolor separately in NQ player_info struct
Simplifies access to the player colors later. Update the sbar usage of
colors to be consistent between NQ/QW as well.

+++ QW/client/client.h
@@ -94,11 +94,10 @@ typedef struct player_info_s {
     byte pl;
 
     // skin information
-    int topcolor;
-    int bottomcolor;
-
-    int _topcolor;
-    int _bottomcolor;
+    byte topcolor;
+    byte bottomcolor;
+    byte _topcolor;
+    byte _bottomcolor;
 
     int spectator;
     byte translations[VID_GRADES * 256];
